Add testing APIs for SuperTextField
