+++ super_editor/lib/src/infrastructure/super_textfield/android/_editing_controls.dart
@@ -32,6 +32,7 @@ class AndroidEditingOverlayControls extends StatefulWidget {
     required this.textContentKey,
     required this.textFieldLayerLink,
     required this.textContentLayerLink,
+    required this.defaultLineHeight,
     required this.handleColor,
     required this.popoverToolbarBuilder,
     this.showDebugPaint = false,
@@ -61,6 +62,10 @@ class AndroidEditingOverlayControls extends StatefulWidget {
   /// [GlobalKey] that references the [SuperSelectableTextState] within
   /// the text field.
   final GlobalKey<SuperSelectableTextState> textContentKey;
+
+  /// The line height to use to paint the caret when there's no text
+  /// in the field to measure.
+  final double defaultLineHeight;
 
   /// The color of the selection handles.
   final Color handleColor;
@@ -129,7 +134,7 @@ class _AndroidEditingOverlayControlsState extends State<AndroidEditingOverlayCon
 
   @override
   void dispose() {
-    widget.editingController.textController.addListener(_rebuildOnNextFrame);
+    widget.editingController.textController.removeListener(_rebuildOnNextFrame);
 
     WidgetsBinding.instance!.removeObserver(this);
 
@@ -515,7 +520,10 @@ class _AndroidEditingOverlayControlsState extends State<AndroidEditingOverlayCon
     _log.finer('Collapsed handle text position: $extentTextPosition');
     final extentHandleOffsetInText = _textPositionToTextOffset(extentTextPosition);
     _log.finer('Collapsed handle text offset: $extentHandleOffsetInText');
-    final extentLineHeight = widget.textContentKey.currentState!.getCharacterBox(extentTextPosition).toRect().height;
+    double extentLineHeight = widget.textContentKey.currentState!.getCharacterBox(extentTextPosition).toRect().height;
+    if (widget.editingController.textController.text.text.isEmpty) {
+      extentLineHeight = widget.defaultLineHeight;
+    }
 
     if (extentHandleOffsetInText == const Offset(0, 0) && extentTextPosition.offset != 0) {
       // The caret offset is (0, 0), but the caret text position isn't at the
@@ -712,6 +720,12 @@ class AndroidEditingOverlayController with ChangeNotifier {
     required this.textController,
     required LayerLink magnifierFocalPoint,
   }) : _magnifierFocalPoint = magnifierFocalPoint;
+
+  @override
+  void dispose() {
+    _handleAutoHideTimer?.cancel();
+    super.dispose();
+  }
 
   bool _isToolbarVisible = false;
   bool get isToolbarVisible => _isToolbarVisible;
