+++ super_editor/lib/src/infrastructure/super_textfield/android/android_textfield.dart
@@ -30,7 +30,7 @@ class SuperAndroidTextfield extends StatefulWidget {
     required this.caretColor,
     required this.selectionColor,
     required this.handlesColor,
-    this.lineHeight,
+    required this.lineHeight,
     this.textInputAction = TextInputAction.done,
     this.popoverToolbarBuilder = _defaultAndroidToolbarBuilder,
     this.showDebugPaint = false,
@@ -103,7 +103,7 @@ class SuperAndroidTextfield extends StatefulWidget {
   /// result in a constantly changing text field height during scrolling.
   /// To avoid that situation, a single, explicit [lineHeight] is
   /// provided and used for all text field height calculations.
-  final double? lineHeight;
+  final double lineHeight;
 
   /// The type of action associated with the action button on the mobile
   /// keyboard.
@@ -223,9 +223,19 @@ class _SuperAndroidTextfieldState extends State<SuperAndroidTextfield> with Sing
   void dispose() {
     _removeEditingOverlayControls();
 
+    WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
+      // Dispose after the current frame so that other widgets have
+      // time to remove their listeners.
+      _editingOverlayController.dispose();
+    });
+
     _textEditingController.removeListener(_onTextOrSelectionChange);
     if (widget.textController == null) {
-      _textEditingController.dispose();
+      WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
+        // Dispose after the current frame so that other widgets have
+        // time to remove their listeners.
+        _textEditingController.dispose();
+      });
     }
 
     _focusNode.removeListener(_onFocusChange);
@@ -288,6 +298,7 @@ class _SuperAndroidTextfieldState extends State<SuperAndroidTextfield> with Sing
           textFieldKey: _textFieldKey,
           textContentLayerLink: _textContentLayerLink,
           textContentKey: _textContentKey,
+          defaultLineHeight: widget.lineHeight,
           handleColor: widget.handlesColor,
           popoverToolbarBuilder: widget.popoverToolbarBuilder,
           showDebugPaint: widget.showDebugPaint,
