+++ super_editor/pubspec.yaml
@@ -19,12 +19,11 @@ dependencies:
   #       then remove this dependency.
   markdown: ^4.0.0
   flutter_lints: ^1.0.0
+  flutter_test:
+    sdk: flutter
   uuid: ^3.0.3
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
   golden_toolkit: ^0.11.0
   mockito: ^5.0.4
